+++ Sources/BaseFoundationiOS/Common/Extensions/SetExtension.swift
@@ -1,0 +1,14 @@
+//
+//  SetExtension.swift
+//
+//
+//  Created by pablo borquez on 10/05/24.
+//
+
+import Foundation
+
+extension Set {
+    public func toArray() -> [Element] {
+        return Array(self)
+    }
+}